aws/AWS-CDK/VPC/vpc-cross-account-peering/requirements.txt
aws-cdk-lib==2.54.0
flake8
boto3
black